* merged from trunk r25704:25717.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/mvm@25718 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ pack.c
@@ -1970,7 +1970,7 @@ pack_unpack(VALUE str, VALUE fmt)
 	    break;
 
 	  case 'P':
-	    if (sizeof(char *) <= send - s) {
+	    if (sizeof(char *) <= (size_t)(send - s)) {
 		VALUE tmp = Qnil;
 		char *t;
 
@@ -2010,7 +2010,7 @@ pack_unpack(VALUE str, VALUE fmt)
 	    if (len > (long)((send - s) / sizeof(char *)))
 		len = (send - s) / sizeof(char *);
 	    while (len-- > 0) {
-		if (send - s < sizeof(char *))
+		if ((size_t)(send - s) < sizeof(char *))
 		    break;
 		else {
 		    VALUE tmp = Qnil;
